Significant revamping to support new InfluxDB client libs, created new function to trim units without regex, which results in a significant speed up 28sec -> 4sec. This is actually catching values better compared to the regex. Although the regex could be modified as well to work more correctly. Both functions are currently used and compared for testing. No longer using slices to store column names and values, instead using a map[string]{}interface since influxdb client.Points uses this. Store time fields as int64 instead of int. Writting to Influxdb is currently commented out, only JSON output to stdout is functional

+++ influxios/structs.go
@@ -6,6 +6,6 @@ package influxios
 type Block struct {
 	Name        string
 	Lines       []string
-	Created     int
-	LastCreated int
+	Created     int64
+	LastCreated int64
 }
